move table importing logic to other modules
