+++ db/tables/operations/select.py
@@ -1,6 +1,8 @@
+import json
 from sqlalchemy import Table, select, join
 
 from db.connection import exec_msar_func
+from db.columns.operations.alter import _transform_column_alter_dict
 from db.deprecated.utils import execute_statement, get_pg_catalog_table
 
 BASE = 'base'
@@ -40,6 +42,22 @@ def get_table_info(schema, conn):
 
 def list_joinable_tables(table_oid, conn, max_depth):
     return exec_msar_func(conn, 'get_joinable_tables', max_depth, table_oid).fetchone()[0]
+
+
+def get_preview(table_oid, column_list, conn, limit=20):
+    """
+    Preview an imported table. Returning the records from the specified columns of the table.
+
+    Args:
+        table_oid: Identity of the imported table in the user's database.
+        column_list: List of settings describing the casts to be applied to the columns.
+        limit: The upper limit for the number of records to return.
+
+    Note that these casts are temporary and do not alter the data in the underlying table,
+    if you wish to alter these settings permanantly for the columns see tables/alter.py.
+    """
+    transformed_column_data = [_transform_column_alter_dict(col) for col in column_list]
+    return exec_msar_func(conn, 'get_preview', table_oid, json.dumps(transformed_column_data), limit).fetchone()[0]
 
 
 def reflect_table(name, schema, engine, metadata, connection_to_use=None, keep_existing=False):
